Fix wrong min-max filter result of parquet reader
