+++ be/src/exec/CMakeLists.txt
@@ -125,6 +125,7 @@ set(EXEC_FILES
     vectorized/sorting/sort_permute.cpp
     vectorized/sorting/compare_column.cpp
     parquet/column_chunk_reader.cpp
+    parquet/column_converter.cpp
     parquet/column_reader.cpp
     parquet/encoding.cpp
     parquet/level_codec.cpp
